requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.3
pydantic==2.5.3
httpx==0.26.0
reportlab==4.0.9
openpyxl==3.1.2
python-dateutil==2.8.2
